Good-Air-main/requirements.txt
apache-airflow==2.10.4
snowflake-connector-python
snowflake-sqlalchemy
apache-airflow-providers-snowflake
apache-airflow-providers-postgres==5.10.1
flask-login
sqlalchemy
unicodecsv
Flask-JWT-Extended
pandas
requests
datetime
scikit-learn
psycopg2-binary==2.9.9



